Cards extracted
Card created
CardsList created
Home edited so it uses CardsList instead of hardcoding
