+++ client/src/components/CardsList/CardsList.scss
@@ -1,0 +1,9 @@
+.MainContainer{
+    display: flex;
+    align-items: center;
+    justify-items: center;
+    justify-content: center;
+    flex-wrap: wrap;
+    text-decoration: none;
+}
+
