+++ client/src/pages/Home/Home.jsx
@@ -1,18 +1,11 @@
 import React from "react";
 import axios from "axios";
-import { Link } from "react-router-dom";
 import "react-circular-progressbar/dist/styles.css";
-import { CircularProgressbarWithChildren } from "react-circular-progressbar";
-import logo from "../../assets/images/trophy.png";
 import "./Home.scss";
-
+import CardsList from "../../components/CardsList/CardsList";
 function Home() {
   // Master
-  let propsvalue = 0;
-  let propsmaxValue = 0;
   const baseURL = "http://localhost:5000/teams";
-  let color = "";
-  let progress;
   const [teams, setTeams] = React.useState(null);
   let teamPoints = [];
   let draw = true;
@@ -36,11 +29,7 @@ function Home() {
       if (response.data != undefined) {
         if (teamPoints[0] !== teamPoints[1])
           draw = false;
-        //console.log(draw);
-        // if (!draw)     // This is for, if we want to and in case of draw, no one have a different style
-        // {}             // Put in this case all the loops of below inside the curly brackets
         arrayTeams.forEach(team => {
-          //console.log("=>" + team.activities[0].point)
           points = 0;
           for(let i = 0; i < team.activities.length; i++) {
             points += team.activities[i].point;
@@ -61,63 +50,7 @@ function Home() {
     });
   }, []);
 
-  if (!teams) return null;
-  const nTeams = teams.length;
-  const cards = [];
-  for (let i = 0; i < nTeams; i++) {
-    let link = `/listTeam/${i + 1}`;
-    propsvalue = 0;
-    if (teams[i].activities.length !== 0) {
-      teams[i].activities.forEach((element) => {
-        propsvalue += element.point;
-      });
-      propsmaxValue = teams[i].activities.length * 15;
-    } else {
-      propsvalue = 0;
-      propsmaxValue = teams[i].activities.length * 15;
-    }
-    progress = (propsvalue / propsmaxValue) * 100;
-    if (progress < 33) color = "red";
-    else if (propsvalue == 0) color = "red";
-    else if (progress < 66) color = "orange";
-    else if (progress > 66) color = "green";
-    cards.push(
-      <Link to={link} className="linkTeamCard">
-        <div className="cardPosition">
-          <h3 className="teamName">{teams[i].name}</h3>
-          {/* Development */}
-          <div className="circular">
-            <CircularProgressbarWithChildren
-              value={propsvalue}
-              maxValue={propsmaxValue}
-              styles={{
-                path: {
-                  stroke: `${color}`,
-                  strokeLinecap: "round",
-                },
-                trail: {
-                  stroke: "#d6d6d6",
-                },
-              }}
-            >
-              <img
-                style={{ width: 30, marginTop: -5 }}
-                src={logo}
-                alt="trophy"
-                className="trophy"
-              />
-              <div style={{ fontSize: 18, marginTop: -5 }} className="words">
-                <strong>{`${propsvalue}/${propsmaxValue}`}</strong>
-              </div>
-            </CircularProgressbarWithChildren>
-          </div>
-          {/* Development */}
-        </div>
-      </Link>
-    );
-  }
-
-  return <div className="MainContainer">{cards}</div>;
+  return <CardsList teams={teams}></CardsList>;
 }
 
 export default Home;
