--- client/src/pages/Home/Home.scss
@@ -1,72 +1,0 @@
-.MainContainer{
-    display: flex;
-    align-items: center;
-    justify-items: center;
-    justify-content: center;
-    flex-wrap: wrap;
-    text-decoration: none;
-}
-
-.linkTeamCard{
-    display: flex;
-    align-items: center;
-    justify-items: center;
-    justify-content: center;
-    flex-direction: column;
-    padding: 0 20px;
-    margin: 40px;
-    background-color: #fff;
-    border-radius: 4px;
-    height: 180px;
-    width: 90px;
-    text-decoration: none;
-    
-    .teamName{
-        padding: 0;
-        margin: -4px 0px 20px 0px;
-        border: 0;
-        font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;//
-        font-size: 27px;
-        text-decoration: none;
-        color: #310863;
-    }
-    .points{
-        padding: 0;
-        margin: 0;
-        border: 0;
-        font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;//
-        text-decoration: none;
-        color: #310863;
-    }
-}
-.cardPosition{
-    display: flex;
-    display: flex;
-    align-items: center;
-    justify-items: center;
-    justify-content: center;
-    flex-direction: column;
-    padding: 0 0 20px 0;
-}
-// Development
-.circular{
-    width: 85;
-    height: 85;
-}
-.trophys{
-    width: 30;
-    margin-top: -5;
-}
-.words{
-    font-size: 18;
-    margin-top: -5;
-    text-decoration: none;
-}
-
-.leader{
-    text-shadow: black 0.1em 0.1em 0.1em;
-    font-size: 27.25px;
-    color : rgb(255, 183, 0);
-    //-webkit-text-stroke-color: 2px black;
-    //background-color: aqua;
-}
